# frontend/src/types/openapi.yaml
...
openapi: 3.0.1
info:
  title: travelPlan API Test
  description: travelPlan의 api 테스트입니다.
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearerAuth: []
...
paths:
...
  "/member/{id}":
    put:
      tags:
      - Member API
      summary: 회원정보 수정
      operationId: updateMember
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/MemberUpdateRequestDto"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/MemberResponseDto"
    delete:
      tags:
      - Member API
      summary: 회원 삭제
      operationId: deleteMember
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
...
  "/admin/notice/{id}":
    get:
      tags:
      - notice-controller
      operationId: getNoticeById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/Notice"
    put:
      tags:
      - notice-controller
      operationId: updateNotice
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/Notice"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/Notice"
    delete:
      tags:
      - notice-controller
      operationId: deleteNotice
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
...
  "/member/refresh":
    post:
      tags:
      - Member API
      summary: 액세스 토큰 재발급
      operationId: refresh
      parameters:
      - name: refreshToken
        in: cookie
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: object
...
  "/member/login":
    post:
      tags:
      - Member API
      summary: 로그인
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/LoginRequestDto"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: object
  "/member/join":
    post:
      tags:
      - Member API
      summary: 회원가입
      operationId: join
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/JoinRequestDto"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/MemberResponseDto"
  "/member/auth/email":
    post:
      tags:
      - Member API
      summary: 이메일 인증
      operationId: validateAuthCode
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/EmailVerificationRequestDto"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
...
  "/admin/notice":
    get:
      tags:
      - notice-controller
      operationId: getAllNotices
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/Notice"
    post:
      tags:
      - notice-controller
      operationId: createNotice
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/Notice"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/Notice"
  "/admin/blacklist":
    get:
      tags:
      - blacklist-controller
      operationId: getBlacklist
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/Blacklist"
    post:
      tags:
      - blacklist-controller
      operationId: addToBlacklist
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/Dto"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
...
  "/member/mypage":
    get:
      tags:
      - Member API
      summary: 마이페이지
      operationId: getMyPage
      parameters:
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/MemberResponseDto"
  "/member/auth/verify-email":
    get:
      tags:
      - Member API
      summary: 이메일 인증 코드 전송
      operationId: requestAuthCode
      parameters:
      - name: email
        in: query
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
...
  "/admin/blacklist/{memberId}":
    delete:
      tags:
      - blacklist-controller
      operationId: removeFromBlacklist
      parameters:
      - name: memberId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
components:
  schemas:
...
    MemberUpdateRequestDto:
      type: object
      properties:
        nickname:
          type: string
        email:
          type: string
        password:
          type: string
        profileImage:
          type: string
        gender:
          type: string
          enum:
          - MALE
          - FEMALE
        ageRange:
          type: string
          enum:
          - TEENS
          - TWENTIES
          - THIRTIES
          - FORTIES_PLUS
        travelStyle:
          type: string
          enum:
          - TOURISM
          - RELAXATION
          - SHOPPING
        aboutMe:
          type: string
    MemberResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        nickname:
          type: string
        profileImage:
          type: string
        gender:
          type: string
          enum:
          - MALE
          - FEMALE
        ageRange:
          type: string
          enum:
          - TEENS
          - TWENTIES
          - THIRTIES
          - FORTIES_PLUS
        travelStyle:
          type: string
          enum:
          - TOURISM
          - RELAXATION
          - SHOPPING
        aboutMe:
          type: string
        rating:
          type: number
          format: double
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        authority:
          type: string
...
    Member:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        password:
          type: string
        nickname:
          type: string
        profileImage:
          type: string
        gender:
          type: string
          enum:
          - MALE
          - FEMALE
        ageRange:
          type: string
          enum:
          - TEENS
          - TWENTIES
          - THIRTIES
          - FORTIES_PLUS
        travelStyle:
          type: string
          enum:
          - TOURISM
          - RELAXATION
          - SHOPPING
        aboutMe:
          type: string
        rating:
          type: number
          format: double
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        authority:
          type: string
        verified:
          type: boolean
    Notice:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        content:
          type: string
        member:
          "$ref": "#/components/schemas/Member"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    LoginRequestDto:
      required:
      - password
      - username
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    JoinRequestDto:
      required:
      - ageRange
      - email
      - gender
      - nickname
      - password
      - travelStyle
      - username
      type: object
      properties:
        username:
          maxLength: 20
          minLength: 4
          pattern: "^[a-zA-Z0-9_]+$"
          type: string
        email:
          type: string
        password:
          maxLength: 2147483647
          minLength: 8
          type: string
        nickname:
          type: string
        profileImage:
          type: string
        gender:
          type: string
          enum:
          - MALE
          - FEMALE
        ageRange:
          type: string
          enum:
          - TEENS
          - TWENTIES
          - THIRTIES
          - FORTIES_PLUS
        travelStyle:
          type: string
          enum:
          - TOURISM
          - RELAXATION
          - SHOPPING
        aboutMe:
          type: string
    EmailVerificationRequestDto:
      required:
      - authCode
      - email
      type: object
      properties:
        email:
          type: string
        authCode:
          maxLength: 6
          minLength: 6
          type: string
    Dto:
      type: object
      properties:
        memberId:
          type: integer
          format: int64
        reason:
          type: string
...
    Blacklist:
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        id:
          type: integer
          format: int64
        member:
          "$ref": "#/components/schemas/Member"
        reason:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      name: Authorization
      in: header
      scheme: bearer
      bearerFormat: JWT